refactor: share string dedupe helper
